[FIX] web_enterprise: action menu is empty on mobile
Before this commit, some buttons action was hidden in css with this rule:
see addons/web/static/src/scss/form_view.scss:128

.o_form_view.o_form_editable .oe_read_only {
    display: none !important;
}
But the form render is only checking for a hidden button with this CSS rule:
button:not(.o_invisible_modifier)
This is why the action menu is not hidden in edit mode.

After this commit, we also remove these buttons in edit mode

Note we have try to fix this bug using this selector:
button:not(.o_invisible_modifier):not(.o_form_view.o_form_editable .oe_read_only)
but at this time the widget is not appended and so the selector is not enable yet.

Steps to reproduce:
* Open Project App
* In the project view, click on "more button" (3 dots) of a project then Edit
* Click on the action menu. The menu is empty (BUG)

Task ID: 2058510

closes odoo/enterprise#5875

Signed-off-by: Adrien Dieudonné (adr) <adr@odoo.com>