[FIX] industry_fsm: do not allow users to change product on done tasks
When an fsm task is marked as done, we generate stock pickings corresponding to the products that were associated to this task.

Before, it was possible to keep adding/removing products after having marked the task as done. This resulted in inconsistencies between what is displayed on the task and what is displayed on the pickings.

In order to solve this we removed the ability to modify product quantities on an fsm task after it is marked as done.

Note that the user can still edit the sale order directly

Task-2077576

closes odoo/enterprise#5963

Signed-off-by: Jérome Maes (jem) <jem@openerp.com>