[FIX] industry_fsm_report: avoid putting time() in xmlid
This practice make the exportation of terms to translate crashing, because
the demo data does not have the same xml id twice (only for the worksheet
template). Removing the `time()` from the xml id does not make a collision
possible in case of re-installation (now all xml are correctly removed).

Thanks to mat for reporting this !
Task-2079099