[ADD] social_demo: introduce a demo module for the social app
This commit introduces a new complete demo module for the social app.

A full module is necessary because we need to override several methods to prevent calls to third
party APIs.

This creates a full sandbox where you can safely play around with posts and comments while checking
the capabilities of our social app.
It will also make it easier to showcase the app to new customers.

Task#2081610

closes odoo/enterprise#6290

Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>