[FIX] industry_fsm: button visibility of 'Mark as done'
The 'Mark as Done' button should be visible in primary, when
- all checked options (from project) are completed ('Extra quotation' is an
exception: completed or not, the primary button should be visible is all other
options are done)
- only 'Extra Quotation' is checked on project (hence, no other real options)
to indicate to the user that he completed all that was expceted.

The secondary button is visible when at least one option is done (except for
'Extra Quote').

So "Extra Quotation" is seen as an obligation when it is the only option checked,
but it not taken into account when there is other options checked.

Task-2074002

X-original-commit: ed2db14e92aa418ca1047c95fa60f006773e208a