[FIX] account_online_sync: split transactions in different statement
Previously we were splitting transactions in the following way:
We only checked the last statement date and based on the split frequency
we decided if we had to create a new statement or not. We did not take into
account the transactions date at all, which means that it was possible to
have daily statement with transactions from another date.

Now we base ourselves on the transactions date and we move each transactions
to the correct statement based on the spit frequency.
-> This implies some changes, if the transactions has to be put in a posted
bank statement, we will add the line inside and reset the statement to 'draft'.
Other transactions will be created in new statements which can possibly be
between two others statements.

closes odoo/enterprise#6116

Signed-off-by: Quentin De Paoli (qdp) <qdp@openerp.com>