[FIX] account_ponto: default value for account.online.journal name
In some case, the description of the account received by Ponto exists in the dict
values but the value is None or an empty character. However this is used for the
name field of account.online.journal which is required. The fix is to move the default
value outside the .get('description').

closes odoo/enterprise#6005

X-original-commit: 36c3c64a38bf4e0f3cf1c92c84076aa3589aac7d
Signed-off-by: Nicolas Martinelli (nim) <nim@odoo.com>