[FIX] l10n_mx_edi: wrong rendering in unit of measure form
Before this commit, after installing l10n_mx_edi a rendering problem was
found in the unit of measure form. This arises because, l10n_mx_edi adds
a new group in an existing group with fields. Note that, a group with
existing fields cannot contain also a group, this will break the
rendering of the aforementioned fields.

Now, the module adds the group outside all the existing ones.

opw-2080722
opw-2082014

closes odoo/enterprise#6138

X-original-commit: 990cb8f389912482c0c733346a55912479ea5ee5
Signed-off-by: Jorge Pinna Puissant (jpp) <jpp@odoo.com>
Co-authored-by: Jorge Pinna Puissant <jpp@odoo.com>