[IMP] social_push_notifications: add warning when testing notifications if blocked
This commit adds a warning message when the user clicks on the "Test Notification" button if his
notifications are turned off or if he's not on an HTTPS environment.

It was confusing before this change because nothing happened when you clicked on the button and it
felt like it was simply not working.

Task#2082478

X-original-commit: 005c679a3f9c4d5cbd60f9e8de2cb830c99b0945