[FIX] industry_fsm: clean demo data
When doing a demo, user will probably take the first tasks in
the FSM pipe, but those ones are marked as done, so all the
buttons are not visible anymore.
To avoid losing the wow effect during a complete flow, we
need to not set thoses tasks as done.

Task-2074002

closes odoo/enterprise#6069

X-original-commit: 7e5b9dc9654133b530d78b4a3fde379982e74982
Signed-off-by: Jérome Maes (jem) <jem@openerp.com>