[FIX] pos_iot: Traceback on close session
When closing a PoS session connected to a payment terminal, we perform
a balance operation and a receipt is printed. The selector used to
retrieve the receipt has been changed is point_of_sale but wasn't
changed here. The receipt was then not found and there was a traceback.
Since the code is now stable the view itself cannot be modified so the
new class is added dynamically.

X-original-commit: 6dd3709a3fb783327d4b2f7c99c3c22881ef929f