[FIX] helpdesk: Wrong datetime format in retrieve_dashboard
The datetime returned by read_group is formated with babel.
Let's consider a ticket with SLA deadline equal to 14/10/2019 13:00 assigned
to a user in French.
So when making a read_group in an other language than en_US,
the group made with 'sla_deadline:hour' was like 13:00 14 oct. (with babel)
instead of 13:00 14 oct (the format expected by datetime with '%I:%M %d %b'
 in function _is_sla_failed).

opw:2082495

closes odoo/enterprise#6041

X-original-commit: 43db985174e20a85b282552a125778bd74fb3db0
Signed-off-by: Simon Goffin (sig) <sig@openerp.com>