[FIX] l10n_mx_edi: Date due empty on invoice duplicated
Avoid that the date due in the invoice is empty when is duplicated.

In the next case:
1. Create a new customer invoice with payment term ``30 Days``
2. Validate it
3. Duplicate the invoice

In this case, in the new invoice, when is validated the date due is
empty.

To fix this, now is called the ``_onchange_invoice_date()`` like Odoo
makes when is assigned the invoice date here
https://github.com/odoo/odoo/blob/13.0/addons/account/models/account_move.py#L2038

closes odoo/enterprise#6057

X-original-commit: fa3f4ea49a854fa98667aaf33617e308ac5b3ef3
Signed-off-by: Nicolas Martinelli (nim) <nim@odoo.com>