[IMP] web_mobile: open app mobile from mail link
Before this commit, when a mail was send, there was any possibility to
open it with the mobile app for all instance of Odoo.

After this commit, all link related to a model will be prefixed with
a custom link. This link is a "FireBase Dynamic Links" site.

Using this allow us, to be able to open the app mobile when it's installed
and redirect to the normal link when the user don't have the app.

Note we force to have absolute url and not relative url to avoid
dynamic-links with relative link.

Ref:
* https://firebase.google.com/docs/dynamic-links/#how_does_it_work

Task ID: 1895451

closes odoo/enterprise#6414

Related: odoo/odoo#39958
Signed-off-by: Adrien Dieudonné (adr) <adr@odoo.com>