[IMP] web_enterprise: Improve settings tabs style on mobile
With this commit, we apply the same style used by the mobile kanban view
to the settings tabs. To avoid duplicated code we create a Javascript mixin
to have the same code for these two cases.

Before this commit, we had to swipe up to the desired tab. It was a bit
boring if you want to select the last one...
Now we can easily scroll between all tabs to be able to tap/click on it.
No need to swipe to see the other ones.

Task ID: 2085899