[FIX] account_asset: Fixed asset account different from depreciation account
It was not possible to select a fixed Asset account that was different from the depreciation account.

From documentation: https://www.accountingtools.com/articles/what-is-accumulated-depreciation.html

It must be possible to specify a Fixed asset account such as "Fixed Assets - Computers & Equipment"
and then a separate depreciation account such as "Accumulated Deprecation - Computers & Equipment"
as a contra-asset account.

opw:2083242

closes odoo/enterprise#6143

X-original-commit: 23b7fe29758fc4d4da82ff6d3d60a85e8b0862ca
Signed-off-by: Simon Goffin (sig) <sig@openerp.com>