[REF] account_reports: Improve performances of the financial reports
====================
Performance analysis
====================

Analysis made on the balance sheet with 5000 journal entries.

                        | time (s)   | # queries
--------------------------------------------------
Before the refactoring:

Load report             | 0.68       | 244
Unfolding line          | 0.19       | 55

After the refactoring:

Load report             | 0.33       | 145
Unfolding line          | 0.16       | 55
--------------------------------------------------

====================
ADDITIONAL POINTS
====================

- Fix the groupby feature allowing some horizontal groupby & make it usable by the user.
This feature is current hidden in the ir.filters. Not really user friendly and leads
to some bugs when unfolding the lines.

- Remove the debit_credit filter.
This feature was used by 2-3 reports. However, since the formulas was simply copy-pasted
and the amounts was only based on the balance sign, we decided to remove this feature.

- Simplify the formulas.
No longer allow a left operand on formula since the debit_credit feature has been removed.
"sum" is now directly the balance, not an object.
E.g:
"balance = sum.balance" is now "sum"

- Allow having multiple headers lines in a more generic way(adapt the trial balance) + remove _get_super_column method.
In some reports, we want multiple headers lines like the financial reports with the additional groupby.
Then, this commit removes such custom stuff allowing this to bring a more generic way to do so.

--task: 2151051