[FIX] mrp_workorder: add button_start in tests
The 'Process' button on workorder open the tablet view but also set the
time the production start effectively. As we simulate it only by openning the
tablet view in tests, we were missing information on the workorder.

This commit call button_start() each time we open the tablet in tests

Task : 2082442

closes odoo/enterprise#6209

X-original-commit: 4d34bd5bae3bd5de26a6b9e14168772e367c09b3
Signed-off-by: Simon Lejeune (sle) <sle@openerp.com>