[IMP] helpdesk: Apply SLA based on ticket tags
Purpose
=======

Tags are useful to flag the various types of tickets handled by a given helpdesk
team. It would be a great help to apply the SLA Policies based on this
information.

Specifications
===========

Add a "Ticket Tags" field below the "Ticket Type" on the SLA Policies form view.
Do not make it mandatory. If no tags are specified, the SLA should be applied on
all tickets.
You should be able to select one or more Helpdesk tags
Helpdesk tags comes as an additional criterion a ticket have to meet to have the
SLA applied.
e.g. for an SLA with ticket_type_id=Foo and tag_ids=Bar,Bou
The ticket must have the same ticket type AND at least both tags.

closes odoo/enterprise#6158

Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>