[FIX] hr_payroll: Don't count Attendances as paid time off
On the payroll reporting screen, attendances are counted as "Paid time off"
in the pie chart.
The reason is the work type is computed based on the boolean field `is_leave` of
a work entry type. But the field can have a NULL value in the database for
work entry types created before the module `hr_payroll` is installed.
This commit changes the SQL view to support NULL values.

closes odoo/enterprise#6180

X-original-commit: 915d27ae4e4dd22d71945fbdc6d596699fdc8ac1
Signed-off-by: lul-odoo <LucasLefevre@users.noreply.github.com>