[FIX] account_sepa: warnings and correct account number for Switzerland
For Switzerland, we sometimes only have the postal account
and the address of the partner to pay.  We need to take the
special l10n_ch_postal field instead of the acc_number in
case of a postal account.

We also removed the warnings in case of Switzerland, because
in almost any Swiss case, we got warnings before that are not
that interesting in the export wizard.

As CAMT was moved to enterprise, add the split for the account
number here.

closes odoo/enterprise#6514

X-original-commit: 643f9d64d8677a1a7bf50ef9ff24867f2899c152
Signed-off-by: Josse Colpaert <jco@openerp.com>