[FIX] documents_account, account_invoice_extract: fix vendor bill creation
Reverts eb6ee7dbfce and b39d9a096e which was a fix of that first commit.
Because automatic sending to the OCR was done only when going through
message_post, they used the fact that _message_post_process_attachments
adds to the thread the attachments on the 'mail.compose.message' model.
However this method also checks the create_uid, which would mean another
horrible hack would be needed to make the feature work for multi-users
setups. Moreover this relies on a feature that is not part of the
interface, private, in another model, and simply undocumented.

But the issue was also caused by the fact that the way documents_account
linked the attachment was non-standard: it bypasses the standard methods
that would have set the message_main_attachment_id by just editing the
related record. So in that case an explicit call to
register_as_main_attachment is needed.
We add automatic sending to this method so that all flows are uniform.
(there are 4 flows, adding attachments in comments or in
the attachment box, programmatically or through the interface.)

Note that we also remove the attachments argument from the call to
message_post since it is useless and misleading.
Similarly in message_post we can avoid the big code duplication by
simply relying on retry_ocr, since at this point the attachments should
already have been processed by the super.

opw 2086239

closes odoo/enterprise#6307

X-original-commit: b5482abebb8c64accbf60467e161dd5ed3197caf
Signed-off-by: Nans Lefebvre (len) <len@odoo.com>