[FIX] account_followup: fix duplicate on follow up line
In the accounting module, there is the possibility to duplicate
payment follow-ups.

If I want to duplicate one of them, I am facing a warning saying
that the day must be unique.

Before this commit:

    - This is impossible to duplicate a payment follow-up

After this commit:

    - The duplication is working (adds 15 days to the maximum delay
      by default)

Since the concerned field is required and has an unique constraint,
this is the best alternative to make the duplication working.

OPW-2086701

closes odoo/enterprise#6269

X-original-commit: 4408a7506506e1f7e61479dd8099cbaedc1792ad
Signed-off-by: Nicolas Martinelli (nim) <nim@odoo.com>