[IMP] industry_fsm: reduce chatter spam for SO
This commit aims at reducing the number of chatter messages when
adjusting the quantity of a product required for a field service task.

Before, each time the quantity of a product was modified, a message was
logged in the chatter. Now it only logs once the product is initially
added.

closes odoo/enterprise#6375

Taskid: 2088929
X-original-commit: cf6d2ddba1f49353c53951a6c145851eaaa1cde2
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>