[FIX] fields: `compute_sudo` defaults to True for stored fields only
The default `compute_sudo=True` makes sense for recomputing stored
fields that are indirectly related to a business operation.  This
ensures that the recomputation of the field does not break an operation
that is not aware of the fields to recompute.

However, computing non-stored fields in superuser mode is usually not
necessary.  It even leads to unexpected values: counting a partner's
sales orders does not give the same result in superuser mode as in
normal mode.  That is why non-stored fields are not computed in
superuser mode by default.

[FIX] sign, mrp_plm, account_asset: adapt the model definition to make
all fields with the same compute method have the same value for
`compute_sudo`.

[FIX] hr_contract_sign: compute field `sign_request_count` in superuser
mode (necessary for access rights issues).

closes odoo/enterprise#6351

X-original-commit: d1bd484f2ca4829210d97b5a000f28dd7ddb3583
Signed-off-by: Olivier Dony (odo) <odo@openerp.com>