[FIX] studio: check monetary field dependency when exporting data
When exporting field data from studio, the dependencies between
fields is considered in order to sort the fields in the resulting
file to ensure a correct importation. However, monetary fields and
their implicit dependency on a res.currency many2one field is not
taken into account. This can lead to an incorrect order of the
fields in the exportation file and therefore errors in the posterior
importation of the data.

opw 2070660

closes odoo/enterprise#6348

X-original-commit: 49e53fb7152f1597e4d6db39c65af648db2d0af7
Signed-off-by: Nicolas Martinelli (nim) <nim@odoo.com>
Co-authored-by: Raphael Collet <rco@odoo.com>