[FIX] currency_rate_live: Using new URL for Banxico currency rates
The SOAP provided by banxico to get the exchange rate now was disabled,
and now must be used REST to get the exchange rate.

Here the Banxico Doc that was used to get the rate:
https://www.banxico.org.mx/SieAPIRest/service/v1/;jsessionid=e330efad50bb0388c61e1fd255d7
https://www.banxico.org.mx/SieAPIRest/service/v1/doc/consultaDatosSerieRango

Considerations: Banxico provides the way to get the tokens to allow get
the exchange rate, and each token has some limits, only could be used
10000 times in the day. By default was generated the token
d03cdee20272f1edc5009a79375f1d942d94acac8348a33245c866831019fef4, but
could be generated a new by each customer and added how system parameter
with the key banxico_token, to avoid conflicts.

Is used the date from today - 1 for the next reason:
In Mexico, Banxico provides the exchange rate regularly to 12:00 p.m.,
but that rate must be used the next day, to have consistence with all
the entries.

Then, if I get the exchange rate now, I need get the exchange rate from
yesterday, because Odoo get the more new rate in the system.

Then is better execute the cron every day to the first hour of the day.

closes odoo/enterprise#6214

closes odoo/enterprise#6284

Extra: Banxico not provide the exchange rate in weekend days.
Original-signed-off-by: Nicolas Martinelli (nim) <nim@odoo.com>
X-original-commit: 362222d3c10fa02dc889292c5429470b4a7e6985
Signed-off-by: Nicolas Martinelli (nim) <nim@odoo.com>