[FIX] l10n_mx_edi: Demo certificates were updated
The demo certificates for the VAT TCM970625MB1 now are not available for
finkok to allow cancel a CFDI.

Now the cancellation process in demo, is send directly to the SAT, and
only support some CSDs, in this case EKU9003173C9 is one of the VAT for
that certificates.

Wiki from finkok = https://wiki.finkok.com/doku.php?id=pruebas
User: integracion
Pass: 1nT36R4c!0N

Extra: The PAC indicated that the CSD content when is send to cancel not
must have ``\n``, this to allow encode in b64, then was replace
that by '' before to send to cancel. (Based on the invoice method)

closes odoo/enterprise#6166

closes odoo/enterprise#6285

Original-signed-off-by: Nicolas Martinelli (nim) <nim@odoo.com>
X-original-commit: f80e05da90dc0076b2fd22d49e7d5daa5d7505a5
Signed-off-by: Nicolas Martinelli (nim) <nim@odoo.com>