[FIX] pos_iot: Manual reading on Adam Scales
At the moment the weight is read from all scales automatically. The
problem is that Adam scales beep every time we read the weight, making
them almost unusable in a shop. We set the manufacturer on the IoT
Device to be able to identify those scales and get the weight manually
instead of reading it automatically.

closes odoo/enterprise#6295

Taskid: 2089995
X-original-commit: 4894721e5d6511541d5d91fe2b05a4259f3473e1
Signed-off-by: pimodoo <pimodoo@users.noreply.github.com>