[FIX] view gantt: Fix unavailability issues on gantt view
Purpose
=======

On a gantt view, let's say for the hr.leave, if an employee is
unavailable from Friday 17pm to Monday 8am, and if the precision
of the gantt view is based on half-days, the employee will be displayed
as unavailable on Friday Afternoon and Monday morning, which is obviously
wrong.

Specification
=============

1/ Client side

Currently, we are considered as unavailable if there is 1 slot during this
period (morning or afternoon).

Now, we compute the sum of all the intervals during the period (morning
or afternoon), and we are considered as unavailable if this sum is greater
than 10 hours. (Threshold chosen empirically, could be modified afterward).

2/ Server side

On the applications in which this feature is used, we were only taking
into account the interval longer than 1 hour for a day-based view, and
1 day for the other periods.

Now, we only take into account the interval longer than 1 hour for the
day/week-based view, and 12 hours for the remaining periods.

closes odoo/enterprise#6586

Taskid: 2088873
X-original-commit: d602df4a5cdc379ed2438f9a74c5565681ad685f
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>