[IMP] social: support @tag and #hashtag links in posts and comments
Done
====
1. In post preview, highlight URL and #hashtag
2. #hashtag now depends on the media
    (#hashtag on Facebook redirects to Facebook,
     before it was redirecting always to Twitter)
3. highlight @tag (Twitter only)

How
===
Rename ``SocialStreamPostFormatterMixin`` to ``SocialPostFormatterMixin``
(because we use it in post preview, comment view, kanban view...)

Add ``SocialEmojisMixin`` into ``SocialPostFormatterMixin``, so the code is cleaner

Add the attribute ``media_type`` on ``SocialPostFormatterMixin``
this attribute is used to determine the URLs we need to use to build anchors

Impacted widgets:
- social_post_preview
- social_kanban_field_stream_post
- social_post_kanban_comments

Facebook case
=============
API
---
The Facebook API returns the position and the length of the tag.
If there is "@tag" in the post, it doesn't means that's a tag... Maybe it's raw text.
We must care about this to parse tag in the message.

How
---
1. Modify posts and comments API calls to get the tags information in the message
2. Replace "simple tag text" by ``@text``
3. Let the widget handle the text rendering

Task #2088851

closes odoo/enterprise#6292

Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>