[FIX] helpdesk_sale: Restrict sale_order_id field to correct group
The `sale_order_id` field on a `helpdesk.ticket` is on the form view.
But if the user does not have the access rights to read sale orders, the field
is empty and useless.
Hence a group attibutes is added on the field definition to match the
`ir.model.access` defined for a sale order.

Task 2070454