[FIX] pos_enterprise: IoT in new PoS config
When you create a new PoS config and you check 'IoT Box', you always
saw the 'Save' button as if pos_iot was not installed yet. The save
button is shown when `is_installed_pos_iot`, a computed field, is set
to False. When creating a new record, this field is not computed yet
so it is considered as False.

We remove the `is_installed_pos` field and simplify the code. The PoS
config view in `pos_enterprise` now overrides the one in `point_of_sale`
and `pos_iot` overrides `pos_enterprise`.

We still need to have the `module_pos_iot` field so that pos_iot will
be installed when saving the PoS config.

closes odoo/enterprise#6323

Taskid: 2090363
Signed-off-by: pimodoo <pimodoo@users.noreply.github.com>