[IMP] account_accountant: add account.reconcile.model.line
Task 2046908
Instead of having all the fields duplicated with second_*, we have now a
o2m allowing us to
* have more than 2 lines
* reduce duplicated code
* fix bugs and add features at only one place

We also remove the computation of writeoff and suggestions from the
client side as some code was 4-upled before (twice in in client and
twice in server side). The logic is now only at one place.

closes odoo/enterprise#6324

Related: odoo/odoo#38119
Signed-off-by: Quentin De Paoli (qdp) <qdp@openerp.com>