[FIX] helpdesk: Don't wait the next hour to update ticket dashboard
Currently, to get the helpdesk dashboard data, tickets are retrieved
using a read_group, with the field `sla_deadline`, aggregated by hour.
A ticket with deadline 8:30 AM falls in a group 8:00 AM -> 9:AM
If it's now 8:45AM, the whole group is not considered to be late, whereas
the ticket should count as failed.