[FIX] account_reports: get AML lines only
The datev export format of the `l10n_de_reports` module makes use of the
`aml_only` context key. By using this context key, it is expected that a
list of AML ids is returned by `_get_lines`.

We add back the support for this context key, since post-filtering the
lines in `l10n_de_reports` implies a more complex processing.

opw-2090156
opw-2089209

closes odoo/enterprise#6371

X-original-commit: 9de01dcb1e50985efe79492b2d36e8e9193e61bc
Signed-off-by: Nicolas Martinelli (nim) <nim@odoo.com>