[FIX] l10n_mx_reports: Avoid multicompany errors
In the next case:
1. Create a company (Company by default) for Mexico
2. Create a vendor bill, with tax 16%
3. Create a new company, also for Mexico
4. Try get the DIOT with the second company

This returns the next traceback:
```
The requested operation ("read" on "Tax" (account.tax)) was rejected
because of the following rules:
- Tax multi-company

Note: this might be a multi-company issue.

(Records: IVA(16%) COMPRAS (id=10), User: Mitchell Admin (id=2))```

The error is because search all the ``account.tax.repartition.line``,
with the tags for DIOT report, but not search only the company records.

Now the domain was improved to get only the records with the same user
company (The DIOT report do not have multicompany options)

closes odoo/enterprise#6372

X-original-commit: 95322ef032220e4b312d52c1f10e4dbf13d56161
Signed-off-by: Nicolas Martinelli (nim) <nim@odoo.com>