[FIX] l10n_mx_edi: Ensure module installation on sh
The hook that download the XSD files for the CFDI was moved to a cron,
this to ensure the module installation on sh instances.

Now not take time to download the XSD files in the module installation.
Was moved for a cron that is executed one time, and could be executed
manually with the same server action that already exists.

closes odoo/enterprise#11056

X-original-commit: cf8fd41c7b2a3d91a7b34fa268bd01b1b44c189c
Signed-off-by: Josse Colpaert <jco@openerp.com>