[IMP] timesheet_grid: improve the access rules for timesheets
TaskID: 2088586

Improvement with this commit :
----------------------------

The user with the access rule : 'See own timesheets' cannot
validate the timesheets.

When the visibility of an project is set on
'Invited employees', the user who isn't Administrator in Timesheets cannot see
the timesheets into the project, if this user doesn't follow the project.

The user with the lowest access right can only see his timesheets and not
all timesheets.

The user cannot update an validated timesheet

Add some tests for access rights

Remove unnecessary rules and add some conditions in write to have a
user-friendly error message when the user may not update a timesheet