[FIX] l10n_mx_edi: Flow to cancel process was fixed
Make consistent cancelling invoice already cancelled on PAC & reset PAC
status when setting invoice to draft state.

Before this commit, invoices that are not signed on SAT (because and
error or missing configuration, etc) and previously
cancelled and set them to draft state are not possible to cancel again
because the method action_invoice_cancel filter the invoices where
l10n_mx_edi_pac_status is on  [False, 'retry', 'to_sign'] only, the rest
of them are not candidates to be cancelled.

So considering the following Case:
    Create an invoice, try to validate it and Fails
    - Invoice is validated
    - PAC status is to_sing or retry

    Call cancellation
    - invoice is cancelled
    - PAC status now is cancelled because of this line
      https://github.com/odoo/enterprise/blob/11.0/l10n_mx_edi/models/account_invoice.py#L580

    Set the invoice to draft
    - Invoice now is on the draft state

    Try to cancel the invoice again
    - it Fails because invoices are filtered by PAC status False, 'retry', 'to_sign'
    and it does not consider invoices in cancelled status.

Also PAC status is reset to False when invoice is set to draft in order
to reset the cycle in order to make consistency.

Fix 2: In the cancel process that reset2open an invoice cancelled in
Odoo and open in the SAT, and the invoices cancelled in the SAT but open
in Odoo, was improved to only consider customer invoices. The vendor
bills cannot be affected with this process.

closes odoo/enterprise#6409

X-original-commit: b1442740fa73e001232314f076c6b9e00cbef01f
Signed-off-by: Nicolas Martinelli (nim) <nim@odoo.com>