[FIX] test_l10n_be_hr_payroll_account : add additional group for holiday manager and payroll user
TaskID: 2088586

The access right for timesheets has changed, now the lowest access right
cannot create timesheets for others employees.

Before the commit, the holiday manager and payroll user into this test
cannot access or create timesheets for others users because the access rights have
changed in the module hr_timesheet and timesheet_grid. In this commit, we
allow these two users to access and create timesheets for others users.