[FIX] documents_product: settings use the current company as fallback
Since the products don't have a default company anymore and following
the refactor of the documents settings system, there was no fallback
for the company on which the settings were obtained.

This commit fixes this issue and allows `product.product` and
`product.template` to generate documents based on the current company
if they don't have a `company_id`.

closes odoo/enterprise#6425

Task: 2092537
X-original-commit: 7f3bd136b136b8f5a9082926bb130c211b61f030
Signed-off-by: Thanh Dodeur <ThanhDodeurOdoo@users.noreply.github.com>