[FIX] account_winbooks_import: multiple file import dropped This commit fixes 3 problems with the same solution: * In winbooks, there seems to be multiple savepoints depending on the local databases of different users. All these savepoints are saved in different filed. Before this commit, we tried to be smart about it and we tried to merge those files.
* It seems you can have plugins for winbooks for some kind of control on
the accounts. The parameters were stored in a file that matched the
pattern we chose to find the files containing the chart of accounts.

* It seems winbooks needs a file for migration of the account
codes/names. This file was also taken in the pattern.

The solution is to not trying to be smart anymore.
It turns out there is not really a heuristic that could make this, so
the import is only based on the "master" files, reducing the possible
patterns.

closes odoo/enterprise#6553

X-original-commit: 5991ec235145fd952186dd9661bc7b12283d018e
Signed-off-by: Quentin De Paoli (qdp) <qdp@openerp.com>