[ADD] sale_coupon_taxcloud: send discounts to taxcloud
TaxCloud does not handle discounts; it does not have any concept for it.
From the documentation: https://taxcloud.com/support/discounts
All discounts must be applied to the Sales Price before calculating tax.
Furthermore, sellers should disclose their methodology to customers
in a Discount Application Policy.

Before, the discounts were never sent to Taxcloud.
We stock them as negative lines, so that we can always remove them and
reapply them when we recompute coupon validity.
However, non-positive lines are not sent to TaxCloud at all.
It follows that the taxes were always computed for the full order.
For example, if the order had a 10% coupon, taxes would be computed for
100% instead of 90%.

There are two problems with the naive solution to send abs(price) for
discounts lines. First, it assumes that the taxes are linearly applied
(i.e. tax(x) + tax(y) = tax(x + y) for all x, y) which is not the case,
but then we still need to know on which line(s) the discount is supposed
to apply, to know the tax category. Which adds another issue with
composite discounts -- fixed amounts or 'on order' apply to multiple
categories.

So we go for the correct solution which is to apply the discounts
directly on the lines before sending them to TaxCloud.
The modeling of this module is as follows: we inject in each created
discount line the coupon program that created it.
This allows us to deduce on which line(s) the discount applies.
Then when we need to send the lines to TaxCloud, after discount.

The algorithm is as follows: line-specific discounts are applied first,
on the lines they are supposed to apply to. This means product-specific
discounts or discounts on the cheapest line.
Then global discounts are applied evenly across paid lines, as much as
possible. If there is a remainder, it is applied sequentially, in the
order of the lines, until there is either no discount left or
no discountable line left.

We want to make sure all discounts are applied, as much as possible.
Since the discounts are actually not linked to a line,
we can make incoherent order, because
each coupon is applied independently of other coupons.

For instance make a 60% discount on the cheapest line, then 50% on order
Then we have two lines, with the second one containing 50% of the
cheapest line and 50% of the rest. So it makes 110% of the cheapest line
and 50% of the rest. If the rest is nothing, we have a 110% discount,
which would then top at 100%. Otherwise, the 10% would be deduced from
the order total.
In the case where we offer 150% of one line, the 150% being all
line-specific, our algorithm will apply the 50% on the rest of the
order.
A test show that in all cases where we all apply some discounts
(including borderline cases) we have a consistency between the order
total and the sum of the "taxcloud prices".

However note that the borderline cases might not be legal in certain
states, since a coupon on product X can end up being applied on products
Y1, ..., Yk.
It is the responsibility of the user to configure discounts correctly;
in any case we stick to existing behaviour of the coupon module.

opw 2079677

X-original-commit: d08e5083fe89df231766bf57606515677405e1ad