[FIX] web_studio: filter out non-existing fields
- Add a related field to a form view

It crashes.

Despite a warning, some random dude changed the behavior by introducing
the following commit in a stable version:

https://github.com/odoo/odoo/commit/9595dd06a1d349b2fb84969374d57c32a5faab74

From this commit, a creation now raises an error on non-existing fields,
which is exactly what happens in this case.

We filter out the fields before creation.

Note that we could have done it here:

https://github.com/odoo/enterprise/blob/b1442740fa73e001232314f076c6b9e00cbef01f/web_studio/static/src/js/views/view_editor_manager.js#L384

However, the fix is more generic at the Python level.

opw-2093196

closes odoo/enterprise#6421

X-original-commit: a81d0eb611278d294a170b6971b3471c710b5090
Signed-off-by: Nicolas Martinelli (nim) <nim@odoo.com>