[ADD] approvals_purchase: new approval type
Technical module to make the link between Approvals & Purchase.
This module adds a new value for `approval_type` in approval category
to be able to generate purchase order.

With a purchase approval the users can create a request for a purchase.
They can enter the product id directly or just a description (the
description is always required because the product asked by the user is
maybe not yet in the DB, or the user don't know what product he/she must
select), then, the approver must check all request lines have a product
id before approve it.
From one approved approval request, a Purchase user can click on the
"Create Purchase Orders" button to create a purchase order or update
quantity of compatible existing purchase order.

task-2040710