[FIX] iot: Terminal blocked by async transactions
To give the possibility to the user to cancel transactions, we used to
execute them in asynchronous mode. This allowed us to listen for abort
actions coming from the PoS. We switched to asynchronous mode before
starting the transaction and back to synchronous right after.

We query the status of the terminal 20 seconds. If the check happened
while the async mode was activated, the call to the API indefinitely
blocked and all the following transactions had to wait. To continue
using his terminal, the user had to reboot his box.

We don't need to execute transactions asynchronously anymore. We listen
for abort actions in the main thread while transactions are processed
in another one.

closes odoo/enterprise#6619

X-original-commit: 1ae71448f6cc81e63bdbacc2e343cb577e429fc2
Signed-off-by: Quentin Lejeune (qle) <qle@odoo.com>